c45: invert port/dev
The MDIO spec uses "port" as the top level primitive on the bus. Each
port can then address 32 "devices". This is the inverse of the
nomenclature the author is used to from working with Marvell switches.

So swap these so that the c45 address format is port:device rather than
the inverse. Also, align mv6tool with these changes.