Fix gcc warnings on unused returns with asprintf
Newer gcc flags for various Linux distros look for unused return
calls. asprintf will return the size of the buffer or -1 if out of
memory. Since this tool is used for hardware enablement where memory
might be tight, it looks good to exit cleanly if memory is not
available versus trying to use a NULL ptr string.

Signed-off-by: Stephen Smoogen <ssmoogen@redhat.com>